another major refactor
* implement a comparison function named `IsEqualEncodingResource` for `EncodingResource` types.

* removed the `DeptIdToDepartment` of type `map[uint16]Departments.Department` in `EncodingResource` type.

* implement function called `GenerateEncodingResourceFromUniTimeTable` that is use to generate an `EncodingResource` instance from a `UniTimeTables` instance.

* the function `EncodeIndividualGenome` now takes a `dept_id_to_department` of type `map[uint16]Departments.Department` instead of getting it inside of the `EncodingResource` type.

* removed section retries

* added a `is_subject_type_added_once` boolean variable in `EncodeIndividualGenome` to avoid double assignment of subjects that has lecture and laboratory classes.

* change the `TotalTeachingHours` property type from `int` to `float32` in the `Instructor` struct type.

* added panic in `IncTimeSlotClassCount` function to avoid overflow of `uint4`

+++ Resources/Instructors/Instructor.go
@@ -12,7 +12,7 @@ type Instructor struct {
 	AssignedSubjects int
 
 	// the total teaching hours assigned for all assigned subjects.
-	TotalTeachingHours int
+	TotalTeachingHours float32
 
 	// AvailableDay, each bits represent the boolean value if the instructor
 	// is available.
